@@ -1,68 +1,70 @@
 {
-	"name": "craue/geo-bundle",
-	"type": "symfony-bundle",
-	"description": "Doctrine functions for calculating geographical distances in your Symfony project.",
-	"keywords": ["geo", "symfony"],
-	"homepage": "https://github.com/craue/CraueGeoBundle",
-	"license": "MIT",
-	"authors": [
-		{
-			"name": "Christian Raue",
-			"email": "christian.raue@gmail.com"
-		},
-		{
-			"name": "Symfony Community",
-			"homepage": "https://github.com/craue/CraueGeoBundle/contributors"
-		}
-	],
-	"require": {
-		"php": "^7.3|^8",
-		"doctrine/doctrine-bundle": "^1.6.12|^2",
-		"doctrine/orm": "^2.5.2",
-		"symfony/config": "~4.4|~5.3|^6",
-		"symfony/dependency-injection": "~4.4|~5.3|^6",
-		"symfony/framework-bundle": "~4.4|~5.3|^6"
-	},
-	"require-dev": {
-		"doctrine/annotations": "^1.6",
-		"doctrine/data-fixtures": "^1.0.1",
-		"doctrine/dbal": "^2.7|^3",
-		"doctrine/doctrine-fixtures-bundle": "^2.4.1|^3.1",
-		"phpstan/extension-installer": "^1.1",
-		"phpstan/phpstan": "^1.4",
-		"phpstan/phpstan-deprecation-rules": "^1.0",
-		"phpstan/phpstan-strict-rules": "^1.1",
-		"phpstan/phpstan-symfony": "^1.1",
-		"phpunit/phpunit": "^9.5",
-		"symfony/browser-kit": "~4.4|~5.3|^6",
-		"symfony/phpunit-bridge": "^6",
-		"symfony/yaml": "~4.4|~5.3|^6"
-	},
-	"suggest": {
-		"doctrine/data-fixtures": "*"
-	},
-	"minimum-stability": "stable",
-	"autoload": {
-		"psr-4": {
-			"Craue\\GeoBundle\\": ""
-		},
-		"exclude-from-classmap": [
-			"/Tests/"
-		]
-	},
-	"extra": {
-		"branch-alias": {
-			"dev-master": "1.9.x-dev"
-		},
-		"symfony": {
-			"require": "~4.4|~5.3|^6"
-		}
-	},
-	"config": {
-		"allow-plugins": {
-			"phpstan/extension-installer": true,
-			"symfony/flex": true
-		},
-		"sort-packages": true
-	}
+  "name": "durimjusaj/geo-bundle",
+  "type": "symfony-bundle",
+  "description": "Doctrine functions for calculating geographical distances in your Symfony project.",
+  "keywords": [
+    "geo",
+    "symfony"
+  ],
+  "homepage": "https://github.com/craue/CraueGeoBundle",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Christian Raue",
+      "email": "christian.raue@gmail.com"
+    },
+    {
+      "name": "Symfony Community",
+      "homepage": "https://github.com/craue/CraueGeoBundle/contributors"
+    }
+  ],
+  "require": {
+    "php": "^8",
+    "doctrine/doctrine-bundle": "^2",
+    "doctrine/orm": "^3",
+    "symfony/config": "^7",
+    "symfony/dependency-injection": "^7",
+    "symfony/framework-bundle": "^7",
+    "symfony/validator": "^7"
+  },
+  "require-dev": {
+    "doctrine/annotations": "^2",
+    "doctrine/data-fixtures": "^2",
+    "doctrine/dbal": "^4",
+    "doctrine/doctrine-fixtures-bundle": "^4",
+    "phpstan/extension-installer": "^1.4",
+    "phpstan/phpstan": "^2.1",
+    "phpstan/phpstan-deprecation-rules": "^2.0",
+    "phpstan/phpstan-strict-rules": "^2.0",
+    "phpstan/phpstan-symfony": "^2.0",
+    "phpunit/phpunit": "^11.5",
+    "symfony/browser-kit": "^7",
+    "symfony/phpunit-bridge": "^7",
+    "symfony/yaml": "^7",
+    "ext-zip": "*"
+  },
+  "suggest": {
+    "doctrine/data-fixtures": "*"
+  },
+  "minimum-stability": "stable",
+  "autoload": {
+    "psr-4": {
+      "Craue\\GeoBundle\\": ""
+    },
+    "exclude-from-classmap": [
+      "/Tests/"
+    ]
+  },
+  "extra": {
+    "symfony": {
+      "require": "7"
+    }
+  },
+  "config": {
+    "allow-plugins": {
+      "phpstan/extension-installer": true,
+      "symfony/flex": true
+    },
+    "sort-packages": true
+  }
 }
