@@ -3,9 +3,10 @@
 namespace Craue\GeoBundle\Doctrine\Query\Mysql;
 
 use Doctrine\ORM\Query\AST\Functions\FunctionNode;
-use Doctrine\ORM\Query\Lexer;
+use Doctrine\ORM\Query\AST\Node;
 use Doctrine\ORM\Query\Parser;
 use Doctrine\ORM\Query\SqlWalker;
+use Doctrine\ORM\Query\TokenType;
 
 /**
  * Usage: GEO_DISTANCE_BY_POSTAL_CODE(countryOrigin, postalCodeOrigin, countryDestination, postalCodeDestination)
@@ -17,22 +18,22 @@ use Doctrine\ORM\Query\SqlWalker;
  */
 class GeoDistanceByPostalCode extends FunctionNode {
 
-	protected $countryOrigin;
-	protected $postalCodeOrigin;
-	protected $countryDestination;
-	protected $postalCodeDestination;
+	protected Node $countryOrigin;
+	protected Node $postalCodeOrigin;
+	protected Node $countryDestination;
+	protected Node $postalCodeDestination;
 
 	public function parse(Parser $parser) : void {
-		$parser->match(Lexer::T_IDENTIFIER);
-		$parser->match(Lexer::T_OPEN_PARENTHESIS);
+		$parser->match(TokenType::T_IDENTIFIER);
+		$parser->match(TokenType::T_OPEN_PARENTHESIS);
 		$this->countryOrigin = $parser->StringPrimary();
-		$parser->match(Lexer::T_COMMA);
+		$parser->match(TokenType::T_COMMA);
 		$this->postalCodeOrigin = $parser->StringPrimary();
-		$parser->match(Lexer::T_COMMA);
+		$parser->match(TokenType::T_COMMA);
 		$this->countryDestination = $parser->StringPrimary();
-		$parser->match(Lexer::T_COMMA);
+		$parser->match(TokenType::T_COMMA);
 		$this->postalCodeDestination = $parser->StringPrimary();
-		$parser->match(Lexer::T_CLOSE_PARENTHESIS);
+		$parser->match(TokenType::T_CLOSE_PARENTHESIS);
 	}
 
 	public function getSql(SqlWalker $sqlWalker) : string {
