@@ -9,71 +9,65 @@ use Symfony\Component\Validator\Constraints as Assert;
  * @copyright 2011-2022 Christian Raue
  * @license http://opensource.org/licenses/mit-license.php MIT License
  */
-class GeoPostalCode {
+class GeoPostalCode
+{
 
-	/**
-	 * @var int
-	 */
-	protected $id;
+    protected int $id;
 
-	/**
-	 * @var string
-	 * @Assert\NotBlank
-	 */
-	protected $country;
+    #[Assert\NotBlank]
+    protected string $country;
 
-	/**
-	 * @var string
-	 * @Assert\NotBlank
-	 */
-	protected $postalCode;
+    #[Assert\NotBlank]
+    protected string $postalCode;
 
-	/**
-	 * @var double
-	 * @Assert\NotBlank
-	 */
-	protected $lat;
+    #[Assert\NotBlank]
+    protected float $lat;
 
-	/**
-	 * @var double
-	 * @Assert\NotBlank
-	 */
-	protected $lng;
+    #[Assert\NotBlank]
+    protected float $lng;
 
-	public function getId() {
-		return $this->id;
-	}
+    public function getId(): int
+    {
+        return $this->id;
+    }
 
-	public function setCountry($country) {
-		$this->country = $country;
-	}
+    public function setCountry(string $country): void
+    {
+        $this->country = $country;
+    }
 
-	public function getCountry() {
-		return $this->country;
-	}
+    public function getCountry(): string
+    {
+        return $this->country;
+    }
 
-	public function setPostalCode($postalCode) {
-		$this->postalCode = $postalCode;
-	}
+    public function setPostalCode(string $postalCode): void
+    {
+        $this->postalCode = $postalCode;
+    }
 
-	public function getPostalCode() {
-		return $this->postalCode;
-	}
+    public function getPostalCode(): string
+    {
+        return $this->postalCode;
+    }
 
-	public function setLat($lat) {
-		$this->lat = $lat;
-	}
+    public function setLat(float $lat): void
+    {
+        $this->lat = $lat;
+    }
 
-	public function getLat() {
-		return $this->lat;
-	}
+    public function getLat(): float
+    {
+        return $this->lat;
+    }
 
-	public function setLng($lng) {
-		$this->lng = $lng;
-	}
+    public function setLng(float $lng): void
+    {
+        $this->lng = $lng;
+    }
 
-	public function getLng() {
-		return $this->lng;
-	}
-
+    public function getLng(): float
+    {
+        return $this->lng;
+    }
 }
